Utils and warnings

+++ GameEngine/GameEngine/Engine/Text/FontManager.cpp
@@ -140,7 +140,7 @@ void FontManager::_draw2DString(const std::string &text,
 	{
 		std::size_t l = std::size_t(text[i] - ASCII_BEGIN);
 		if (l > ASCII_END)
-			glyphWidth = size;
+			glyphWidth = static_cast<float>(size);
 		else
 			glyphWidth = f._size != size ? ((float)f._map[l].width / (float)f._size) * sz : f._map[l].width;
 		if (text[i] == ' ')
